pipeline/requirements.txt
sqlalchemy>=2.0
psycopg2-binary>=2.9
pandas>=2.0
boto3>=1.34
opensearch-py>=2.4
graphrag-toolkit>=3.16
click>=8.1
python-dotenv>=1.0
